Scoring improvements:  * Scores now track % economy and combat.  * Scores now track damage done and taken.  * Biomass history no longer shows frame 0 twice.  * Biomass history now repeats the last sample after the end of the round. This makes scores containing rounds that ended early (by elimination) less misleading - now they show a sudden cessation of growth instead of dropping to zero.

+++ source/support/GBLongNumber.h
@@ -23,7 +23,7 @@ public:
 	GBLongNumber & operator+=(const GBNumber addend);
 	GBLongNumber & operator+=(const GBLongNumber & addend);
 // misc
-	long operator/(const GBLongNumber & divisor) const;
+	double operator/(const GBLongNumber & divisor) const;
 	long operator/(long divisor) const;
 	bool Nonzero() const;
 	bool Zero() const;
